feat: mover FerramentasSection para o diretório de componentes e adicionar dados de ferramentas
